+++ src/next.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const config = {
+    extends: ['@idearium/eslint-config/src/common', 'next', 'prettier'],
+};
+
+module.exports = config;
